build: Fetch defconfig according to ARCH and BOARD.
This patch brings two features to the build system.

The major feature added is that defconfig are stored at
kern-ucore/arch/$(ARCH)/configs and are fetched according to the variable ARCH
and BOARD given, e.g. "make ARCH=arm BOARD=goldfishv7 menuconfig".

The another feature is that, after .config has been set up, no ARCH or BOARD are
needed any longer. The simple "make kernel" or "make sfsimg" will do the work as
ARCH will be fetched from .config and BOARD is not used in the build process.

In short, It is possible to build the system using the following commands now:
        make ARCH=arm BOARD=goldfishv7 menuconfig
        make sfsimg
        make kernel

Signed-off-by: Junjie Mao <eternal.n08@gmail.com>