autotest: Support platform-specific tests by @arch
If @arch is present in a test spec, the script will test if the current arch is
required to pass the test by finding $UCONFIG_ARCH in the value of @arch.

e.g. the following line
    @arch i386
means this test is only required on uCore for i386 and will be ignored when we
are testing uCore on other platforms.

Signed-off-by: Junjie Mao <eternal.n08@gmail.com>