Provide a config entry for minimizing initial files in sfsimg
The original initial files occupy more than 20M (most of which are
minigui-related binaries and resources and should be cleaned up in the future).
These files have run over the size limit of external storage on many embedded
systems but are totally unnecessary.

This patch provides a config entry at 'Filesystem/Minimize contents'. When
selected, only some manuals are included in addition to the utilities.

The config entry is default to n. It should be explicitly set to y if needed.

Signed-off-by: Junjie Mao <eternal.n08@gmail.com>