Add a single-threaded version on x86_64 called amd64
This patch provides a single-threaded implementation for x86_64 so that
multi-core support can be introduced without considering the original
version. This is the first step to get rid of the supervisor as well as
the dirty glue_* files.

Not that this version does not support swapping yet. As swapping is
planned to be removed, this should not be a problem.

Signed-off-by: Junjie Mao <eternal.n08@gmail.com>