Fix sfs_dirtest1
The test list entries under various directories and expects the 'home'
directory to be empty. However, there must be at least one file under
that directory so that git never ignore it. Thus the number of entries
in 'home' should be 1 (. and .. are not counted) instead of 0 and its
size (number of entries include . and .. multiplies size of the name
buffer, i.e. 256) should be 768 instead of 512.

Signed-off-by: Junjie Mao <eternal.n08@gmail.com>