Fix brktest
The total size brktest allocates may vary on different architectures
as there is a pointer field in the slot which is 4-byte on i386 and
arm and 8-byte on amd64.

This patch saves the pointers to the slots allocated in a specific
array and removes the pointer in struct slot so that the result will
be the same on all architectures.

Signed-off-by: Junjie Mao <eternal.n08@gmail.com>