Fix faultreadkernel
Issue #4

In faultreadkernel, the process attempts to read a word from kernel
space and fails if it can successfully do so.

The address previously used is specific to 64-bit implementations and
the testspec is specific to x86 architectures. This patch fixes the
problem by reading different address according to the size of a pointer
and remove the architecture-related lines in the specification.

Signed-off-by: Junjie Mao <eternal.n08@gmail.com>